Avoid NM circular dependency, set default log level
The NormalMeans class had a circular dependency: the Prior class calls
the NormalMeans class, while the NormalMeans requires the Prior class
for initialization. Earlier, it was implemented using a @classmethod
which created the NormalMeans class based on the Prior input. I changed
it to a function to call the NormalMeans class. The CustomerLogger class
can now set default log level for the whole program without requiring
specification from child classes (unless required).